Improved login-styling and moved a few files
- Refactored login_screen to use TealGradLeft and TealGradRight
- Added background and better textcolor for 'Forgot password' in login_screen
- Refactored home_screen to use TealGradLeft and TealGradRight
- Moved textsection and titlesection to home/components
- Deleted loginpanel widget
