+++ app/lib/screens/login/login_screen.dart
@@ -1,3 +1,6 @@
+import 'package:app/components/tealgradleft.dart';
+import 'package:app/components/tealgradright.dart';
+import 'package:app/components/topbar.dart';
 import 'package:app/screens/login/components/button.dart';
 import 'package:flutter/material.dart';
 import 'package:app/screens/login/components/textfield.dart';
@@ -21,70 +24,76 @@ class _LoginPageState extends State<LoginPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.grey[300],
-      body:  SafeArea(
-        child: Center(
-          child: Column(
-            children: [
-              const SizedBox(height: 20),
+      appBar: const TopBar(),
+      body: Stack (
+        children: [
+          const TealGradLeft(),
+          const TealGradRight(),
+          SafeArea(
+            child: Center(
+              child: Column(
+                children: [
+                  const SizedBox(height: 20),
 
-              const Text(
-                'Credentials',
-                style: TextStyle(
-                  fontSize: 20,
-                ),),
-              const SizedBox(height: 20),
+                  const Text(
+                    'Credentials',
+                    style: TextStyle(
+                      fontSize: 20,
+                    ),),
+                  const SizedBox(height: 20),
 
-              // username textfield
-              CustomTextField(
-                controller: usernameController,
-                hintText: 'Username',
-                obscureText: false,
-              ),
+                  // username textfield
+                  CustomTextField(
+                    controller: usernameController,
+                    hintText: 'Username',
+                    obscureText: false,
+                  ),
 
-              const SizedBox(height: 10),
+                  const SizedBox(height: 10),
 
-              // password textfield
-              CustomTextField(
-                controller: passwordController,
-                hintText: 'Password',
-                obscureText: true,
-              ),
+                  // password textfield
+                  CustomTextField(
+                    controller: passwordController,
+                    hintText: 'Password',
+                    obscureText: true,
+                  ),
 
-              const SizedBox(height: 10),
+                  const SizedBox(height: 10),
 
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 25.0),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: [
-                    TextButton(
-                      onPressed: forgotPassword,
-                      child: Text(
-                      'Forgot Password?',
-                      style: TextStyle(color: Colors.grey[600]),
+                  Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 25.0),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        TextButton(
+                          onPressed: forgotPassword,
+                          child: const Text(
+                            'Forgot Password?',
+                            style: TextStyle(color: Colors.white),
+                          ),
+                        ),
+                      ],
                     ),
-                    ),
-                  ],
-                ),
+                  ),
+
+                  const SizedBox(height: 10),
+
+                  CustomButton(
+                    onTap: loginUser),
+                  
+                  const SizedBox(height: 30),
+                  
+                  Divider(),
+
+                  Text('or use the QR code'),
+
+                  Icon(Icons.qr_code_2_rounded, size: 180),
+                ],
               ),
-
-              const SizedBox(height: 10),
-
-              CustomButton(
-                onTap: loginUser),
-              
-              const SizedBox(height: 30),
-              
-              Divider(),
-
-              Text('or use the QR code'),
-
-              Icon(Icons.qr_code_2_rounded, size: 180),
-            ],
+            ),
           ),
-        ),
-      ),
+        ],
+      )  
     );
   }
 }
